+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [11.2.0] 2026-05-04
+### Added
+- Exported Questionnaire JSON now follows FHIR canonical field order for all
+  backbone elements (Questionnaire, Item, EnableWhen, AnswerOption, Initial).
+
 ## [11.1.0] 2026-04-23
 ### Added
 - Added support to create general FHIR extensions including nested extensions.